complex HTTP server using Node's HTTP module
server is assigned to the variable app that is exported HTTP server listening on port 1245 and returns plain text, when the URL path is /, it should display Hello Holberton School! in the page body. When the URL path is /students, it should display: This is the list of our students followed by the same content as the file 3-read_file_async.js (with and without the database) - the name of the database must be passed as argument of the file
CSV file can contain empty lines (at the end) - and they are not a valid student!